% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-req.r
\name{curl_convert}
\alias{curl_convert}
\title{Shortcut to convert a single cURL command-line into a single R function}
\usage{
curl_convert()
}
\value{
an R function and a version of the function on the clipboard
}
\description{
Shortcut to convert a single cURL command-line into a single R function
}
\note{
The cURL command-line should be on the clipboard.
}
