% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from-har.r
\name{from_request}
\alias{from_request}
\title{Convert HAR request to cURL}
\usage{
from_request(request)
}
\arguments{
\item{request}{a parsed HAR entry  request as a list}
}
\value{
a cURL string
}
\description{
Convert HAR request to cURL
}
\examples{
\dontrun{
   hardir <- system.file("extdata", package = "curlconverter")
   harfile <- file.path(hardir, "har1.har")
   entry <-
     fromJSON(harfile, simplifyVector = FALSE)[["log"]][["entries"]][[1]]
   req <- entry[["request"]]
   req \%>\% from_request
}
}
