% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_req.r
\name{make_req}
\alias{make_req}
\title{Turn a cURL request into an \code{httr} request}
\usage{
make_req(curls = read_clip(), quiet = FALSE)
}
\arguments{
\item{curls}{a character vector of one cURL command line. It will
read from the clipboard (i.e. if you did a \emph{"Copy as cURL"} from
browser developer tools).}

\item{quiet}{if \code{FALSE}, a \code{message} with the original \code{cURL}
command line will be output and the created \code{httr} function will
be \code{cat}ted to the console.. (Default: \code{FALSE})}
}
\value{
character vector containing a \code{httr} request
}
\description{
Takes in a \emph{"Copy as cURL"} command line and returns source code
for an \code{httr} \code{VERB} request via the clipboard, function return value
and (optionally) \code{cat}ted to the console.
}

