% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_query.r
\name{parse_query}
\alias{parse_query}
\title{Split a query string into component parts}
\usage{
parse_query(query)
}
\arguments{
\item{query}{query string to decode}
}
\description{
While \link{straighten} will returns parsed \code{GET} query string
parameters there are times (i.e. in HTML \code{<form>} processing) when
the body of the request contains a URL encoded query string as well.
}
\details{
This function will take any query string and return a named list of
the paremters. Both the names and values will be URL decoded.
}
\examples{
parse_query("a=1&b=test")
}
\references{
\href{https://en.wikipedia.org/wiki/Query_string}{Query Strings}
}

