% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curlconverter-package.r
\docType{package}
\name{curlconverter}
\alias{curlconverter}
\alias{curlconverter-package}
\title{Create `httr` requests from _"Copy to cURL"_ command-line calls}
\description{
Create `httr` requests from _"Copy to cURL"_ command-line calls
}
\author{
Bob Rudis (@hrbrmstr)
}

