% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_har.r
\name{from_har}
\alias{from_har}
\title{Convert HAR to cURL}
\usage{
from_har(har, filter = c("All", "Doc", "JS", "CSS", "XHR", "Img", "XHR",
  "Media", "Font", "WS", "Manifest", "Other"))
}
\arguments{
\item{har}{a JSON string, URL or file}

\item{filter}{a filter working on content-type given by the HAR entry
response}
}
\value{
A character vector of cURL calls
}
\description{
Convert HAR to cURL
}
\examples{
\dontrun{
   hardir <- system.file("extdata", package = "curlconverter")
   harfile <- file.path(hardir, "har1.har")
   harfile \%>\% from_har
}
}

