% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_req.r
\name{make_req}
\alias{make_req}
\title{Turn a parsed cURL requests into \code{httr} request functions}
\usage{
make_req(x, quiet = TRUE)
}
\arguments{
\item{x}{a vector of \code{curlcoverter} objects}

\item{quiet}{if \code{FALSE}, will cause \code{make_req()} to write complete function
source code to the console.}
}
\value{
a \code{list} of working R \code{function}s.
}
\description{
Takes the output of \code{straighten()} and turns the parsed cURL command lines
into working \code{httr} functions, optionally \code{cat}'ing the text of each function
to the console.
}
\examples{
\dontrun{
library(httr)
library(magrittr)
 my_ip <- straighten("curl 'https://httpbin.org/ip'") \%>\% make_req()
 content(my_ip[[1]](), as="parsed")
}
}

