% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_req.r
\name{make_req}
\alias{make_req}
\title{Turn a cURL request into an \code{httr} request}
\usage{
make_req(curls = read_clip(), quiet = FALSE)
}
\arguments{
\item{curls}{a character vector of one cURL command line. It will
read from the clipboard (i.e. if you did a \emph{"Copy as cURL"} from
browser developer tools).}

\item{quiet}{if \code{FALSE}, a \code{message} with the original \code{cURL}
command line will be output and the created \code{httr} function will
be \code{cat}ted to the console.. (Default: \code{FALSE})}
}
\value{
a working R \code{function}
}
\description{
Takes in a \emph{"Copy as cURL"} command line and returns a "working"
R \code{function} (i.e. a comple, valid \code{VERB()} call that will work
provided the server it's requesting the resource from thinks the call is
valid).
}
\details{
It also has a side-effect of overwriting the clipboard with the character
contents of the function and will (optionally) \code{cat}ted to the console.
}
\examples{
\dontrun{
 my_ip <- make_req("curl 'https://httpbin.org/ip'")
 content(my_ip(), as="parsed")
}
}

