% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curlconverter.r
\name{straighten}
\alias{straighten}
\title{Processes new cURL request}
\usage{
straighten(curls = read_clip())
}
\arguments{
\item{curls}{a character vector of one or more cURL command lines. It will read from the
clipboard (i.e. if you did a \emph{"Copy as cURL"} from browser developer tools).}
}
\value{
list of data to use in `httr` requests
}
\description{
Will eventually have an option to return a function, or function source
Right now it just returns daa
}

